requirements.txt
asgiref==3.5.2
Django==4.1.2
django-extensions==3.2.1
djangorestframework==3.14.0
numpy==1.23.4
plotly==5.6.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tenacity==8.1.0
torch==1.12.0
typing_extensions==4.4.0
